+++ morpha/morpha/models/__init__.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 
 from .irecord import IRecord
+from .damage_record import DamageRecord
 from .team_preview_record import TeamPreviewRecord
 from .battle_log import BattleLog
 
 __all__ = ['BattleLog',
+           'DamageRecord',
            'IRecord',
            'TeamPreviewRecord']